refactor: split HipToLLVM.cpp monolith into per-operator files
Split the 2500-line HipToLLVM.cpp into 15 per-operator lowering files
and a shared HipToLLVMUtils.h header. This eliminates the single-file
bottleneck that causes merge conflicts when multiple developers add
or modify operators concurrently.

Each operator's LLVM lowering pattern is now self-contained in its own
translation unit, with a populate*Patterns() function that the pass
calls to register patterns. The shared header provides inline helper
functions (extractMemRefPtr, computeNumElements, etc.) and runtime
function name constants.

No behavioral changes - pure mechanical code movement.
