+++ lib/Conversion/HipToLLVM/ActivationLowering.cpp
@@ -1,0 +1,166 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "HipToLLVMUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+// hip.sigmoid(ctx, input, output)
+//   -> wrap_miopenActivationForward(state, input, output, num_elements,
+//                                    data_type, activation_mode=0)
+// Supports both static and dynamic shapes (computes num_elements at runtime).
+struct SigmoidOpLowering : public ConvertOpToLLVMPattern<SigmoidOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(SigmoidOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type ptrType = getPtrType();
+    Type i32Type = rewriter.getI32Type();
+    Type i64Type = rewriter.getI64Type();
+
+    // Helper to create i64 constants
+    auto createI64Const = [&](int64_t value) -> Value {
+      return LLVM::ConstantOp::create(rewriter, loc, i64Type,
+                                      rewriter.getI64IntegerAttr(value));
+    };
+
+    // Extract pointers using alignedPtr (respects memref.view offsets)
+
+    Value statePtr = adaptor.getCtx();
+    Value inputPtr = extractMemRefPtr(adaptor.getInput(), rewriter, loc);
+    Value outputPtr = extractMemRefPtr(adaptor.getOutput(), rewriter, loc);
+
+    auto outputType = cast<MemRefType>(op.getOutput().getType());
+
+    // Compute num_elements (supports dynamic shapes)
+    // Start with constant 1, multiply by each dimension (static or dynamic)
+    Value numElements = createI64Const(1);
+    MemRefDescriptor outputDesc(adaptor.getOutput());
+
+    for (auto dimIdx : llvm::seq<int64_t>(outputType.getRank())) {
+      Value dimSize;
+      if (outputType.isDynamicDim(dimIdx)) {
+        // Dynamic dimension: extract from runtime descriptor
+        dimSize = outputDesc.size(rewriter, loc, dimIdx);
+      } else {
+        // Static dimension: use compile-time constant
+        dimSize = createI64Const(outputType.getDimSize(dimIdx));
+      }
+      numElements = LLVM::MulOp::create(rewriter, loc, numElements, dimSize);
+    }
+
+    // Get data type enum (f32=0, f16=1, bf16=2)
+    int64_t dataType = getHipdnnDataType(outputType.getElementType());
+    if (dataType < 0)
+      return rewriter.notifyMatchFailure(
+          op, "unsupported element type for hip.sigmoid");
+
+    Value dataTypeVal = createI64Const(dataType);
+    Value activationModeVal = createI64Const(0); // HIPDNN_EP_ACTIVATION_SIGMOID
+
+    // int wrap_miopenActivationForward(RuntimeState* state, void* input,
+    //     void* output, int64_t num_elements, int64_t data_type,
+    //     int64_t activation_mode)
+    SmallVector<Type, 6> paramTypes = {ptrType, ptrType, ptrType,
+                                       i64Type, i64Type, i64Type};
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kWrapMiopenActivationForward, paramTypes, i32Type);
+    if (failed(funcOp))
+      return failure();
+
+    SmallVector<Value, 6> args = {statePtr,    inputPtr,    outputPtr,
+                                  numElements, dataTypeVal, activationModeVal};
+
+    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+// hip.silu(handle, input, output)
+struct SiluOpLowering : public ConvertOpToLLVMPattern<SiluOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(SiluOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type voidType = getVoidType();
+    Type ptrType = getPtrType();
+
+    SmallVector<Type> paramTypes(3, ptrType);
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipSilu, paramTypes, voidType);
+    if (failed(funcOp))
+      return failure();
+
+    SmallVector<Value> args = {
+        adaptor.getCtx(), extractMemRefPtr(adaptor.getInput(), rewriter, loc),
+        extractMemRefPtr(adaptor.getOutput(), rewriter, loc)};
+
+    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+// hip.miopen.softmax(%handle) ins(%input) outs(%output)
+//   -> hip_miopen_softmax(handle, input, output, rows, cols)
+// Rank-generic: softmax over last dim. For 3D [B,S,D], rows = B*S, cols = D.
+struct MiopenSoftmaxOpLowering
+    : public ConvertOpToLLVMPattern<MiopenSoftmaxOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(MiopenSoftmaxOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type voidType = getVoidType();
+    Type ptrType = getPtrType();
+    Type indexType = getIndexType();
+
+    SmallVector<Type> paramTypes = {ptrType, ptrType, ptrType, indexType,
+                                    indexType};
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kMiopenSoftmax, paramTypes, voidType);
+    if (failed(funcOp))
+      return failure();
+
+    int rank = cast<MemRefType>(op.getInput().getType()).getRank();
+    MemRefDescriptor inputDesc(adaptor.getInput());
+
+    // cols = last dim; rows = product of all other dims
+    Value cols = inputDesc.size(rewriter, loc, rank - 1);
+    Value rows = inputDesc.size(rewriter, loc, 0);
+    for (int i = 1; i < rank - 1; i++)
+      rows = LLVM::MulOp::create(rewriter, loc, rows,
+                                 inputDesc.size(rewriter, loc, i));
+
+    SmallVector<Value> args = {
+        adaptor.getCtx(), extractMemRefPtr(adaptor.getInput(), rewriter, loc),
+        extractMemRefPtr(adaptor.getOutput(), rewriter, loc), rows, cols};
+
+    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateActivationLoweringPatterns(const LLVMTypeConverter &converter, RewritePatternSet &patterns) {
+  patterns.add<SigmoidOpLowering, SiluOpLowering, MiopenSoftmaxOpLowering>(converter);
+}
+
+} // namespace hip
+} // namespace mlir
