+++ lib/Conversion/HipToLLVM/ConvLowering.cpp
@@ -1,0 +1,205 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "HipToLLVMUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+// ===== Convolution ops ================================
+
+// hip.conv(%ctx, %input, %weights, %bias, %output)
+//   -> wrap_miopenConvolutionForward(ctx, input, input_n, input_c, input_h,
+//                                     input_w, weights, weights_k, bias,
+//                                     output, output_h, output_w, kernel_h,
+//                                     kernel_w, stride_h, stride_w, pad_top,
+//                                     pad_left, pad_bottom, pad_right,
+//                                     dilation_h, dilation_w, group)
+struct ConvOpLowering : public ConvertOpToLLVMPattern<ConvOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(ConvOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type ptrType = getPtrType();
+    Type i32Type = rewriter.getI32Type();
+    Type i64Type = rewriter.getI64Type();
+
+    // Generate call to runtime wrapper following opaque RuntimeState pattern.
+    // The wrapper extracts handle/stream from state internally (no direct field
+    // access!).
+    //
+    // Signature:
+    // int wrap_miopenConvolutionForward(
+    //     RuntimeState* state,    // Opaque pointer - extracts handle/stream
+    //                             // internally
+    //     void* input,            // Input tensor data pointer
+    //     int64_t input_n,        // Input batch size
+    //     int64_t input_c,        // Input channels
+    //     int64_t input_h,        // Input height
+    //     int64_t input_w,        // Input width
+    //     void* weights,          // Weights tensor data pointer
+    //     int64_t weights_k,      // Output channels (number of filters)
+    //     void* bias,             // Bias tensor data pointer (nullable)
+    //     void* output,           // Output tensor data pointer (in-place)
+    //     int64_t output_h,       // Output height
+    //     int64_t output_w,       // Output width
+    //     int64_t kernel_h,       // Kernel height
+    //     int64_t kernel_w,       // Kernel width
+    //     int64_t stride_h,       // Stride height
+    //     int64_t stride_w,       // Stride width
+    //     int64_t pad_top,        // Padding top
+    //     int64_t pad_left,       // Padding left
+    //     int64_t pad_bottom,     // Padding bottom
+    //     int64_t pad_right,      // Padding right
+    //     int64_t dilation_h,     // Dilation height
+    //     int64_t dilation_w,     // Dilation width
+    //     int64_t group           // Number of groups
+    // );
+    //
+    // Returns: 0 on success, non-zero on error
+
+    // Helper to create i64 constants
+    auto createI64Const = [&](int64_t value) -> Value {
+      return LLVM::ConstantOp::create(rewriter, loc, i64Type,
+                                      rewriter.getI64IntegerAttr(value));
+    };
+
+    // Extract memref pointers (aligned pointers from descriptors)
+    Value statePtr = adaptor.getCtx(); // RuntimeState* (opaque)
+    Value inputPtr = extractMemRefPtr(adaptor.getInput(), rewriter, loc);
+    Value weightsPtr = extractMemRefPtr(adaptor.getWeights(), rewriter, loc);
+    Value outputPtr = extractMemRefPtr(adaptor.getOutput(), rewriter, loc);
+
+    // Handle optional bias
+    Value biasPtr;
+    if (adaptor.getBias()) {
+      biasPtr = extractMemRefPtr(adaptor.getBias(), rewriter, loc);
+    } else {
+      // Pass null pointer if no bias
+      biasPtr = LLVM::ZeroOp::create(rewriter, loc, ptrType);
+    }
+
+    // Extract shapes from memref types
+    // Supports both static and dynamic dimensions using MemRefDescriptor
+    auto inputType = cast<MemRefType>(op.getInput().getType());
+    auto weightsType = cast<MemRefType>(op.getWeights().getType());
+    auto outputType = cast<MemRefType>(op.getOutput().getType());
+
+    // Verify ranks
+    if (inputType.getRank() != 4) {
+      return op.emitError("Input must be rank-4 tensor [N, C, H, W]");
+    }
+    if (weightsType.getRank() != 4) {
+      return op.emitError("Weights must be rank-4 tensor [K, C, R, S]");
+    }
+    if (outputType.getRank() != 4) {
+      return op.emitError("Output must be rank-4 tensor [N, K, H', W']");
+    }
+
+    // Input shape: [N, C, H, W]
+    Value inputN =
+        getMemRefDimSize(inputType, 0, adaptor.getInput(), rewriter, loc);
+    Value inputC =
+        getMemRefDimSize(inputType, 1, adaptor.getInput(), rewriter, loc);
+    Value inputH =
+        getMemRefDimSize(inputType, 2, adaptor.getInput(), rewriter, loc);
+    Value inputW =
+        getMemRefDimSize(inputType, 3, adaptor.getInput(), rewriter, loc);
+
+    // Weights shape: [K, C, R, S] where K=output channels
+    Value weightsK =
+        getMemRefDimSize(weightsType, 0, adaptor.getWeights(), rewriter, loc);
+
+    // Output shape: [N, K, H', W']
+    Value outputH =
+        getMemRefDimSize(outputType, 2, adaptor.getOutput(), rewriter, loc);
+    Value outputW =
+        getMemRefDimSize(outputType, 3, adaptor.getOutput(), rewriter, loc);
+
+    // Extract attributes
+    auto kernelShape = op.getKernelShape();
+    auto strides = op.getStrides();
+    auto pads = op.getPads();
+    auto dilations = op.getDilations();
+    auto group = op.getGroup();
+
+    // Extract integer values from attributes
+    auto getI64 = [](Attribute attr) -> int64_t {
+      return cast<IntegerAttr>(attr).getInt();
+    };
+
+    Value kernelH = createI64Const(getI64(kernelShape[0]));
+    Value kernelW = createI64Const(getI64(kernelShape[1]));
+    Value strideH = createI64Const(getI64(strides[0]));
+    Value strideW = createI64Const(getI64(strides[1]));
+    Value padTop = createI64Const(getI64(pads[0]));
+    Value padLeft = createI64Const(getI64(pads[1]));
+    Value padBottom = createI64Const(getI64(pads[2]));
+    Value padRight = createI64Const(getI64(pads[3]));
+    Value dilationH = createI64Const(getI64(dilations[0]));
+    Value dilationW = createI64Const(getI64(dilations[1]));
+    Value groupVal = createI64Const(group);
+
+    // Build function signature
+    SmallVector<Type, 24> paramTypes = {
+        ptrType, // state
+        ptrType, // input
+        i64Type, // input_n
+        i64Type, // input_c
+        i64Type, // input_h
+        i64Type, // input_w
+        ptrType, // weights
+        i64Type, // weights_k
+        ptrType, // bias
+        ptrType, // output
+        i64Type, // output_h
+        i64Type, // output_w
+        i64Type, // kernel_h
+        i64Type, // kernel_w
+        i64Type, // stride_h
+        i64Type, // stride_w
+        i64Type, // pad_top
+        i64Type, // pad_left
+        i64Type, // pad_bottom
+        i64Type, // pad_right
+        i64Type, // dilation_h
+        i64Type, // dilation_w
+        i64Type  // group
+    };
+
+    // Lookup or create the runtime function
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kMiopenConvolutionForward, paramTypes, i32Type);
+    if (failed(funcOp))
+      return failure();
+
+    // Build argument list matching the signature
+    SmallVector<Value, 24> args = {
+        statePtr,   inputPtr, inputN,    inputC,    inputH,  inputW,
+        weightsPtr, weightsK, biasPtr,   outputPtr, outputH, outputW,
+        kernelH,    kernelW,  strideH,   strideW,   padTop,  padLeft,
+        padBottom,  padRight, dilationH, dilationW, groupVal};
+
+    // Call the runtime function
+    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+
+    // Erase the HIP conv operation (it's in-place, no results)
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateConvLoweringPatterns(const LLVMTypeConverter &converter, RewritePatternSet &patterns) {
+  patterns.add<ConvOpLowering>(converter);
+}
+
+} // namespace hip
+} // namespace mlir
