+++ lib/Conversion/HipToLLVM/MatmulLowering.cpp
@@ -1,0 +1,113 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "HipToLLVMUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+// ===== hipBLASLt ops =========================================================
+
+// hip.hipblaslt.matmul(handle) ins(A, B) outs(C)
+//   -> hip_hipblaslt_matmul(handle, A, B, C, rankA, rankB, batch, M, K, N)
+// Rank-generic: batch from A if 3D, B broadcast if rankB < rankA.
+struct MatmulOpLowering : public ConvertOpToLLVMPattern<MatmulOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(MatmulOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type ptrType = getPtrType();
+    Type i32Type = rewriter.getI32Type();
+    Type i64Type = rewriter.getI64Type();
+
+    // Helper: create i64 constant
+    auto createI64Const = [&](int64_t value) -> Value {
+      return LLVM::ConstantOp::create(rewriter, loc, i64Type,
+                                      rewriter.getI64IntegerAttr(value));
+    };
+
+    // Extract pointers
+    Value statePtr = adaptor.getCtx();
+    Value APtr = extractMemRefPtr(adaptor.getA(), rewriter, loc);
+    Value BPtr = extractMemRefPtr(adaptor.getB(), rewriter, loc);
+    Value outputPtr = extractMemRefPtr(adaptor.getOutput(), rewriter, loc);
+
+    // Get memref types and shapes
+    auto AType = cast<MemRefType>(op.getA().getType());
+    auto BType = cast<MemRefType>(op.getB().getType());
+    int64_t ARank = AType.getRank();
+    int64_t BRank = BType.getRank();
+
+    // === DYNAMIC SHAPE SUPPORT ===
+    // For dynamic shapes, we compute dimensions at runtime
+    MemRefDescriptor ADesc(adaptor.getA());
+    MemRefDescriptor BDesc(adaptor.getB());
+
+    // Compute M, K, N from runtime dimensions
+    // A: [..., M, K], B: [..., K, N] or B: [K, N]
+    Value M =
+        (ARank >= 2) ? ADesc.size(rewriter, loc, ARank - 2) : createI64Const(1);
+    Value K = ADesc.size(rewriter, loc, ARank - 1);
+    Value N = BDesc.size(rewriter, loc, BRank - 1);
+
+    // Compute batch count from leading dimensions of A
+    Value batchCount;
+    if (ARank == 2) {
+      batchCount = createI64Const(1);
+    } else {
+      batchCount = ADesc.size(rewriter, loc, 0);
+      for (int64_t i = 1; i < ARank - 2; ++i) {
+        Value dim = ADesc.size(rewriter, loc, i);
+        batchCount = LLVM::MulOp::create(rewriter, loc, batchCount, dim);
+      }
+    }
+
+    // Compute element size in bytes
+    unsigned elemBits = AType.getElementType().getIntOrFloatBitWidth();
+    Value elemSize = createI64Const(elemBits / 8);
+
+    // Runtime signature:
+    // int wrap_hipblasLtMatmul(RuntimeState* state,
+    //                          const void* A, const void* B, void* output,
+    //                          int64_t M, int64_t N, int64_t K,
+    //                          int64_t batch_count, int64_t elem_size)
+    SmallVector<Type, 9> paramTypes = {
+        ptrType, // state
+        ptrType, // A
+        ptrType, // B
+        ptrType, // output
+        i64Type, // M
+        i64Type, // N
+        i64Type, // K
+        i64Type, // batch_count
+        i64Type  // elem_size
+    };
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kWrapHipblasltMatmul, paramTypes, i32Type);
+    if (failed(funcOp))
+      return failure();
+
+    SmallVector<Value, 9> args = {statePtr, APtr, BPtr,       outputPtr, M,
+                                  N,        K,    batchCount, elemSize};
+
+    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateMatmulLoweringPatterns(const LLVMTypeConverter &converter, RewritePatternSet &patterns) {
+  patterns.add<MatmulOpLowering>(converter);
+}
+
+} // namespace hip
+} // namespace mlir
