+++ lib/Conversion/HipToLLVM/MemoryLowering.cpp
@@ -1,0 +1,290 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "HipToLLVMUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+// --- AllocOp: hip.alloc(%ctx, %dyn...) -> hipMalloc(bytes) + memref descriptor
+struct AllocOpLowering : public ConvertOpToLLVMPattern<AllocOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(AllocOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    MemRefType memRefType = op.getMemref().getType();
+
+    if (!isConvertibleAndHasIdentityMaps(memRefType))
+      return rewriter.notifyMatchFailure(op, "incompatible memref type");
+
+    // Declare hipMalloc(size: i64) -> ptr
+    Type indexType = getIndexType();
+    Type ptrType = getPtrType();
+    FailureOr<LLVM::LLVMFuncOp> mallocFn = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipMalloc, indexType, ptrType);
+    if (failed(mallocFn))
+      return failure();
+
+    // Compute sizes and sizeBytes (dynamic sizes are after the ctx).
+    SmallVector<Value, 4> sizes;
+    SmallVector<Value, 4> strides;
+    Value sizeBytes;
+    getMemRefDescriptorSizes(loc, memRefType, adaptor.getDynamicSizes(),
+                             rewriter, sizes, strides, sizeBytes, true);
+
+    Value allocatedPtr =
+        LLVM::CallOp::create(rewriter, loc, *mallocFn, sizeBytes).getResult();
+
+    // Cast to memref address space if needed
+    Type elementPtrType = getElementPtrType(memRefType);
+    if (!elementPtrType)
+      return rewriter.notifyMatchFailure(op,
+                                         "could not compute element ptr type");
+    FailureOr<unsigned> addrSpace =
+        getTypeConverter()->getMemRefAddressSpace(memRefType);
+    if (failed(addrSpace))
+      return failure();
+    if (cast<LLVM::LLVMPointerType>(allocatedPtr.getType()).getAddressSpace() !=
+        *addrSpace)
+      allocatedPtr = LLVM::AddrSpaceCastOp::create(
+          rewriter, loc,
+          LLVM::LLVMPointerType::get(rewriter.getContext(), *addrSpace),
+          allocatedPtr);
+
+    MemRefDescriptor desc = createMemRefDescriptor(
+        loc, memRefType, allocatedPtr, allocatedPtr, sizes, strides, rewriter);
+    rewriter.replaceOp(op, {desc});
+    return success();
+  }
+};
+
+// --- FreeOp: hip.free(%ctx, %memref) -> llvm.call @hipFree(allocated_ptr)
+struct FreeOpLowering : public ConvertOpToLLVMPattern<FreeOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(FreeOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type voidType = getVoidType();
+    Type ptrType = getPtrType();
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp =
+        LLVM::lookupOrCreateFn(rewriter, module, kHipFree, ptrType, voidType);
+    if (failed(funcOp))
+      return failure();
+
+    Value memrefDesc = adaptor.getMemref();
+    Value allocatedPtr =
+        MemRefDescriptor(memrefDesc).allocatedPtr(rewriter, loc);
+    // hipFree expects void*; if memref is in non-default address space, cast
+    auto ptrTy = allocatedPtr.getType();
+    if (cast<LLVM::LLVMPointerType>(ptrTy).getAddressSpace() != 0)
+      allocatedPtr =
+          LLVM::AddrSpaceCastOp::create(rewriter, loc, ptrType, allocatedPtr);
+
+    LLVM::CallOp::create(rewriter, loc, *funcOp, allocatedPtr);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+// --- GetPoolOp: hip.get_pool(%ctx, %pool_size) : memref<?xi8>
+//     -> llvm.call @hipdnn_ep_get_pool_base(state, size) + memref descriptor
+struct GetPoolOpLowering : public ConvertOpToLLVMPattern<GetPoolOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(GetPoolOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type ptrType = getPtrType();
+    Type i64Type = rewriter.getI64Type();
+    MemRefType memRefType = cast<MemRefType>(op.getPool().getType());
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipGetPoolBase, {ptrType, i64Type}, ptrType);
+    if (failed(funcOp))
+      return failure();
+
+    Value poolSize = adaptor.getPoolSize();
+    Value rawPtr = LLVM::CallOp::create(rewriter, loc, *funcOp,
+                                        ValueRange{adaptor.getCtx(), poolSize})
+                       .getResult();
+
+    FailureOr<unsigned> addrSpace =
+        getTypeConverter()->getMemRefAddressSpace(memRefType);
+    if (failed(addrSpace))
+      return failure();
+
+    Value gpuPtr = rawPtr;
+    if (cast<LLVM::LLVMPointerType>(rawPtr.getType()).getAddressSpace() !=
+        *addrSpace)
+      gpuPtr = LLVM::AddrSpaceCastOp::create(
+          rewriter, loc,
+          LLVM::LLVMPointerType::get(rewriter.getContext(), *addrSpace),
+          rawPtr);
+
+    Value stride1 = LLVM::ConstantOp::create(
+        rewriter, loc, rewriter.getI64Type(), rewriter.getI64IntegerAttr(1));
+
+    MemRefDescriptor desc = createMemRefDescriptor(
+        loc, memRefType, gpuPtr, gpuPtr, {poolSize}, {stride1}, rewriter);
+    rewriter.replaceOp(op, {desc});
+    return success();
+  }
+};
+
+//===----------------------------------------------------------------------===//
+// GetConstantOp Lowering
+//===----------------------------------------------------------------------===//
+
+struct GetConstantOpLowering : public ConvertOpToLLVMPattern<GetConstantOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(GetConstantOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type ptrType = LLVM::LLVMPointerType::get(rewriter.getContext(), 0);
+    Type i64Type = IntegerType::get(rewriter.getContext(), 64);
+    MemRefType memRefType = op.getResult().getType();
+
+    SmallVector<Type, 2> paramTypes = {ptrType, i64Type};
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipGetConstant, paramTypes, ptrType);
+    if (failed(funcOp))
+      return failure();
+
+    SmallVector<Value, 2> args = {adaptor.getCtx(), adaptor.getIndex()};
+    auto callOp = LLVM::CallOp::create(rewriter, loc, *funcOp, args);
+
+    // The runtime always returns a generic pointer (AS 0).  Cast to the
+    // memref's address space (e.g. AS 1 = AMDGPU global memory) if needed.
+    FailureOr<unsigned> addrSpace =
+        getTypeConverter()->getMemRefAddressSpace(memRefType);
+    if (failed(addrSpace))
+      return failure();
+
+    Value dataPtr = callOp.getResult();
+    if (*addrSpace != 0)
+      dataPtr = LLVM::AddrSpaceCastOp::create(
+          rewriter, loc,
+          LLVM::LLVMPointerType::get(rewriter.getContext(), *addrSpace),
+          dataPtr);
+
+    auto shape = memRefType.getShape();
+    SmallVector<Value, 4> sizes;
+    SmallVector<Value, 4> strides;
+
+    for (int64_t dim : shape) {
+      Value size = LLVM::ConstantOp::create(rewriter, loc, i64Type,
+                                            rewriter.getI64IntegerAttr(dim));
+      sizes.push_back(size);
+    }
+
+    int64_t stride = 1;
+    for (int i = shape.size() - 1; i >= 0; --i) {
+      Value strideVal = LLVM::ConstantOp::create(
+          rewriter, loc, i64Type, rewriter.getI64IntegerAttr(stride));
+      strides.insert(strides.begin(), strideVal);
+      stride *= shape[i];
+    }
+
+    MemRefDescriptor desc = createMemRefDescriptor(
+        loc, memRefType, dataPtr, dataPtr, sizes, strides, rewriter);
+
+    rewriter.replaceOp(op, {desc});
+    return success();
+  }
+};
+
+// --- memref.alloc -> hip_device_malloc (produced by bufferization for
+// tensor.empty)
+struct MemRefAllocOpLowering : public ConvertOpToLLVMPattern<memref::AllocOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(memref::AllocOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    MemRefType memRefType = op.getType();
+
+    if (!isConvertibleAndHasIdentityMaps(memRefType))
+      return rewriter.notifyMatchFailure(op, "incompatible memref type");
+
+    Type indexType = getIndexType();
+    Type ptrType = LLVM::LLVMPointerType::get(rewriter.getContext(), 0);
+    FailureOr<LLVM::LLVMFuncOp> mallocFn = LLVM::lookupOrCreateFn(
+        rewriter, module, kHipMalloc, indexType, ptrType);
+    if (failed(mallocFn))
+      return failure();
+
+    SmallVector<Value, 4> sizes;
+    SmallVector<Value, 4> strides;
+    Value sizeBytes;
+    getMemRefDescriptorSizes(loc, memRefType, adaptor.getDynamicSizes(),
+                             rewriter, sizes, strides, sizeBytes, true);
+
+    Value allocatedPtr =
+        LLVM::CallOp::create(rewriter, loc, *mallocFn, sizeBytes).getResult();
+
+    MemRefDescriptor desc = createMemRefDescriptor(
+        loc, memRefType, allocatedPtr, allocatedPtr, sizes, strides, rewriter);
+    rewriter.replaceOp(op, {desc});
+    return success();
+  }
+};
+
+// --- memref.dealloc -> hip_device_free (produced by bufferization)
+struct MemRefDeallocOpLowering
+    : public ConvertOpToLLVMPattern<memref::DeallocOp> {
+  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
+
+  LogicalResult
+  matchAndRewrite(memref::DeallocOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    Location loc = op.getLoc();
+    ModuleOp module = op->getParentOfType<ModuleOp>();
+    Type voidType = LLVM::LLVMVoidType::get(rewriter.getContext());
+    Type ptrType = LLVM::LLVMPointerType::get(rewriter.getContext(), 0);
+
+    FailureOr<LLVM::LLVMFuncOp> funcOp =
+        LLVM::lookupOrCreateFn(rewriter, module, kHipFree, ptrType, voidType);
+    if (failed(funcOp))
+      return failure();
+
+    // Must use allocatedPtr (not alignedPtr) -- hipFree requires the original
+    // allocation base.  With memref.view, alignedPtr points into the pool
+    // interior while allocatedPtr is the pool base.
+    Value allocatedPtr =
+        MemRefDescriptor(adaptor.getMemref()).allocatedPtr(rewriter, loc);
+    if (cast<LLVM::LLVMPointerType>(allocatedPtr.getType()).getAddressSpace() !=
+        0)
+      allocatedPtr =
+          LLVM::AddrSpaceCastOp::create(rewriter, loc, ptrType, allocatedPtr);
+    LLVM::CallOp::create(rewriter, loc, *funcOp, allocatedPtr);
+    rewriter.eraseOp(op);
+    return success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateMemoryLoweringPatterns(const LLVMTypeConverter &converter, RewritePatternSet &patterns) {
+  patterns.add<AllocOpLowering, FreeOpLowering, GetPoolOpLowering, GetConstantOpLowering, MemRefAllocOpLowering, MemRefDeallocOpLowering>(converter);
+}
+
+} // namespace hip
+} // namespace mlir
